+++ qos/cosmos/response_cometbft_status.go
@@ -3,11 +3,12 @@ package cosmos
 import (
 	"encoding/json"
 
+	"github.com/pokt-network/poktroll/pkg/polylog"
+
 	"github.com/buildwithgrove/path/gateway"
 	qosobservations "github.com/buildwithgrove/path/observation/qos"
 	"github.com/buildwithgrove/path/qos"
 	"github.com/buildwithgrove/path/qos/jsonrpc"
-	"github.com/pokt-network/poktroll/pkg/polylog"
 )
 
 // TODO_IMPROVE(@commoddity): Replace custom structs with official CometBFT types.
